more tewaks for seek on > 2Gb files on 32-bit platforms.
git-svn-id: https://svn.r-project.org/R/trunk@31166 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/gnuwin32/run.c
@@ -426,7 +426,7 @@ static int Wpipe_fgetc(Rconnection con)
 }
 
 
-static long null_seek(Rconnection con, double where, int origin, int rw)
+static double null_seek(Rconnection con, double where, int origin, int rw)
 {
     error("seek not enabled for this connection");
     return 0; /* -Wall */
